Struct google_dfareporting2d4::AdvertiserMethods

pub struct AdvertiserMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on advertiser resources. It is not used directly, but through the Dfareporting hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_dfareporting2d4 as dfareporting2d4;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dfareporting2d4::Dfareporting;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dfareporting::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.advertisers();

Methods

impl<'a, C, A> AdvertiserMethods<'a, C, A>

fn insert(&self, request: Advertiser, profile_id: &str) -> AdvertiserInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts a new advertiser.

Arguments

request - No description provided.

profileId - User profile ID associated with this request.

fn patch(&self, request: Advertiser, profile_id: &str, id: &str) -> AdvertiserPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing advertiser. This method supports patch semantics.

Arguments

request - No description provided.

profileId - User profile ID associated with this request.

id - Advertiser ID.

fn list(&self, profile_id: &str) -> AdvertiserListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of advertisers, possibly filtered.

Arguments

profileId - User profile ID associated with this request.

fn update(&self, request: Advertiser, profile_id: &str) -> AdvertiserUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing advertiser.

Arguments

request - No description provided.

profileId - User profile ID associated with this request.

fn get(&self, profile_id: &str, id: &str) -> AdvertiserGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets one advertiser by ID.

Arguments

profileId - User profile ID associated with this request.

id - Advertiser ID.