Struct google_dfareporting2d4::DeliverySchedule

pub struct DeliverySchedule {
    pub priority: Option<String>,
    pub hard_cutoff: Option<bool>,
    pub frequency_cap: Option<FrequencyCap>,
    pub impression_ratio: Option<String>,
}

Delivery Schedule.

This type is not used in any activity, and only used as part of another schema.

Fields

priority

Serving priority of an ad, with respect to other ads. The lower the priority number, the greater the priority with which it is served.

hard_cutoff

Whether or not hard cutoff is enabled. If true, the ad will not serve after the end date and time. Otherwise the ad will continue to be served until it has reached its delivery goals.

frequency_cap

Limit on the number of times an individual user can be served the ad within a specified period of time.

impression_ratio

Impression ratio for this ad. This ratio determines how often each ad is served relative to the others. For example, if ad A has an impression ratio of 1 and ad B has an impression ratio of 3, then DCM will serve ad B three times as often as ad A. Must be between 1 and 10.