Struct google_dfareporting2d4::Dfareporting

pub struct Dfareporting<C, A> {
    // some fields omitted
}

Central instance to access all Dfareporting related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_dfareporting2d4 as dfareporting2d4;
use dfareporting2d4::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dfareporting2d4::Dfareporting;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dfareporting::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.reports().list("profileId")
             .sort_order("accusam")
             .sort_field("takimata")
             .scope("justo")
             .page_token("amet.")
             .max_results(-81)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Methods

impl<'a, C, A> Dfareporting<C, A> where C: BorrowMut<Client>, A: GetToken

fn new(client: C, authenticator: A) -> Dfareporting<C, A>

fn account_active_ad_summaries(&'a self) -> AccountActiveAdSummaryMethods<'a, C, A>

fn account_permission_groups(&'a self) -> AccountPermissionGroupMethods<'a, C, A>

fn account_permissions(&'a self) -> AccountPermissionMethods<'a, C, A>

fn account_user_profiles(&'a self) -> AccountUserProfileMethods<'a, C, A>

fn accounts(&'a self) -> AccountMethods<'a, C, A>

fn ads(&'a self) -> AdMethods<'a, C, A>

fn advertiser_groups(&'a self) -> AdvertiserGroupMethods<'a, C, A>

fn advertisers(&'a self) -> AdvertiserMethods<'a, C, A>

fn browsers(&'a self) -> BrowserMethods<'a, C, A>

fn campaign_creative_associations(&'a self) -> CampaignCreativeAssociationMethods<'a, C, A>

fn campaigns(&'a self) -> CampaignMethods<'a, C, A>

fn change_logs(&'a self) -> ChangeLogMethods<'a, C, A>

fn cities(&'a self) -> CityMethods<'a, C, A>

fn connection_types(&'a self) -> ConnectionTypeMethods<'a, C, A>

fn content_categories(&'a self) -> ContentCategoryMethods<'a, C, A>

fn countries(&'a self) -> CountryMethods<'a, C, A>

fn creative_assets(&'a self) -> CreativeAssetMethods<'a, C, A>

fn creative_field_values(&'a self) -> CreativeFieldValueMethods<'a, C, A>

fn creative_fields(&'a self) -> CreativeFieldMethods<'a, C, A>

fn creative_groups(&'a self) -> CreativeGroupMethods<'a, C, A>

fn creatives(&'a self) -> CreativeMethods<'a, C, A>

fn dimension_values(&'a self) -> DimensionValueMethods<'a, C, A>

fn directory_site_contacts(&'a self) -> DirectorySiteContactMethods<'a, C, A>

fn directory_sites(&'a self) -> DirectorySiteMethods<'a, C, A>

fn event_tags(&'a self) -> EventTagMethods<'a, C, A>

fn files(&'a self) -> FileMethods<'a, C, A>

fn floodlight_activities(&'a self) -> FloodlightActivityMethods<'a, C, A>

fn floodlight_activity_groups(&'a self) -> FloodlightActivityGroupMethods<'a, C, A>

fn floodlight_configurations(&'a self) -> FloodlightConfigurationMethods<'a, C, A>

fn inventory_items(&'a self) -> InventoryItemMethods<'a, C, A>

fn landing_pages(&'a self) -> LandingPageMethods<'a, C, A>

fn metros(&'a self) -> MetroMethods<'a, C, A>

fn mobile_carriers(&'a self) -> MobileCarrierMethods<'a, C, A>

fn operating_system_versions(&'a self) -> OperatingSystemVersionMethods<'a, C, A>

fn operating_systems(&'a self) -> OperatingSystemMethods<'a, C, A>

fn order_documents(&'a self) -> OrderDocumentMethods<'a, C, A>

fn orders(&'a self) -> OrderMethods<'a, C, A>

fn placement_groups(&'a self) -> PlacementGroupMethods<'a, C, A>

fn placement_strategies(&'a self) -> PlacementStrategyMethods<'a, C, A>

fn placements(&'a self) -> PlacementMethods<'a, C, A>

fn platform_types(&'a self) -> PlatformTypeMethods<'a, C, A>

fn postal_codes(&'a self) -> PostalCodeMethods<'a, C, A>

fn projects(&'a self) -> ProjectMethods<'a, C, A>

fn regions(&'a self) -> RegionMethods<'a, C, A>

fn remarketing_list_shares(&'a self) -> RemarketingListShareMethods<'a, C, A>

fn remarketing_lists(&'a self) -> RemarketingListMethods<'a, C, A>

fn reports(&'a self) -> ReportMethods<'a, C, A>

fn sites(&'a self) -> SiteMethods<'a, C, A>

fn sizes(&'a self) -> SizeMethods<'a, C, A>

fn subaccounts(&'a self) -> SubaccountMethods<'a, C, A>

fn targetable_remarketing_lists(&'a self) -> TargetableRemarketingListMethods<'a, C, A>

fn user_profiles(&'a self) -> UserProfileMethods<'a, C, A>

fn user_role_permission_groups(&'a self) -> UserRolePermissionGroupMethods<'a, C, A>

fn user_role_permissions(&'a self) -> UserRolePermissionMethods<'a, C, A>

fn user_roles(&'a self) -> UserRoleMethods<'a, C, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/1.0.0.

Returns the previously set user-agent.