Struct google_dfareporting2d4::DirectorySiteContact

pub struct DirectorySiteContact {
    pub kind: Option<String>,
    pub first_name: Option<String>,
    pub title: Option<String>,
    pub last_name: Option<String>,
    pub id: Option<String>,
    pub phone: Option<String>,
    pub role: Option<String>,
    pub address: Option<String>,
    pub type_: Option<String>,
    pub email: Option<String>,
}

Contains properties of a Site Directory contact.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get directory site contacts (response)

list directory site contacts (none)

Fields
kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#directorySiteContact".

first_name: Option<String>

First name of this directory site contact.

title: Option<String>

Title or designation of this directory site contact.

last_name: Option<String>

Last name of this directory site contact.

id: Option<String>

ID of this directory site contact. This is a read-only, auto-generated field.

phone: Option<String>

Phone number of this directory site contact.

role: Option<String>

Directory site contact role.

address: Option<String>

Address of this directory site contact.

type_: Option<String>

Directory site contact type.

email: Option<String>

Email address of this directory site contact.