Struct google_dfareporting2d4::EventTag

pub struct EventTag {
    pub status: Option<String>,
    pub site_ids: Option<Vec<String>>,
    pub campaign_id: Option<String>,
    pub campaign_id_dimension_value: Option<DimensionValue>,
    pub enabled_by_default: Option<bool>,
    pub advertiser_id: Option<String>,
    pub site_filter_type: Option<String>,
    pub type_: Option<String>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub url: Option<String>,
    pub ssl_compliant: Option<bool>,
    pub exclude_from_adx_requests: Option<bool>,
    pub url_escape_levels: Option<i32>,
}

Contains properties of an event tag.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete event tags (none)

patch event tags (request|response)

list event tags (none)

update event tags (request|response)

insert event tags (request|response)

get event tags (response)

Fields
status: Option<String>

Status of this event tag. Must be ENABLED for this event tag to fire. This is a required field.

site_ids: Option<Vec<String>>

Filter list of site IDs associated with this event tag. The siteFilterType determines whether this is a whitelist or blacklist filter.

campaign_id: Option<String>

Campaign ID of this event tag. This field or the advertiserId field is required on insertion.

campaign_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the campaign. This is a read-only, auto-generated field.

enabled_by_default: Option<bool>

Whether this event tag should be automatically enabled for all of the advertiser's campaigns and ads.

advertiser_id: Option<String>

Advertiser ID of this event tag. This field or the campaignId field is required on insertion.

site_filter_type: Option<String>

Site filter type for this event tag. If no type is specified then the event tag will be applied to all sites.

type_: Option<String>

Event tag type. Can be used to specify whether to use a third-party pixel, a third-party JavaScript URL, or a third-party click-through URL for either impression or click tracking. This is a required field.

advertiser_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the advertiser. This is a read-only, auto-generated field.

id: Option<String>

ID of this event tag. This is a read-only, auto-generated field.

account_id: Option<String>

Account ID of this event tag. This is a read-only field that can be left blank.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#eventTag".

subaccount_id: Option<String>

Subaccount ID of this event tag. This is a read-only field that can be left blank.

name: Option<String>

Name of this event tag. This is a required field and must be less than 256 characters long.

url: Option<String>

Payload URL for this event tag. The URL on a click-through event tag should have a landing page URL appended to the end of it. This field is required on insertion.

ssl_compliant: Option<bool>

Whether this tag is SSL-compliant or not. This is a read-only field.

exclude_from_adx_requests: Option<bool>

Whether to remove this event tag from ads that are trafficked through DoubleClick Bid Manager to Ad Exchange. This may be useful if the event tag uses a pixel that is unapproved for Ad Exchange bids on one or more networks, such as the Google Display Network.

url_escape_levels: Option<i32>

Number of times the landing page URL should be URL-escaped before being appended to the click-through event tag URL. Only applies to click-through event tags as specified by the event tag type.