Struct google_dfareporting2d4::OrderDocument

pub struct OrderDocument {
    pub order_id: Option<String>,
    pub last_sent_time: Option<String>,
    pub amended_order_document_id: Option<String>,
    pub last_sent_recipients: Option<Vec<String>>,
    pub approved_by_user_profile_ids: Option<Vec<String>>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub advertiser_id: Option<String>,
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub effective_date: Option<String>,
    pub title: Option<String>,
    pub project_id: Option<String>,
    pub signed: Option<bool>,
    pub created_info: Option<LastModifiedInfo>,
    pub cancelled: Option<bool>,
    pub type_: Option<String>,
}

Contains properties of a DoubleClick Planning order document.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get order documents (response)

list order documents (none)

Fields

order_id

ID of the order from which this order document is created.

last_sent_time

Timestamp of the last email sent with this order document.

amended_order_document_id

The amended order document ID of this order document. An order document can be created by optionally amending another order document so that the change history can be preserved.

last_sent_recipients

List of email addresses that received the last sent document.

approved_by_user_profile_ids

IDs of users who have approved this order document.

id

ID of this order document.

account_id

Account ID of this order document.

advertiser_id

Advertiser ID of this order document.

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#orderDocument".

subaccount_id

Subaccount ID of this order document.

effective_date

Effective date of this order document.

title

Title of this order document.

project_id

Project ID of this order document.

signed

Whether this order document has been signed.

created_info

Information about the creation of this order document.

cancelled

Whether this order document is cancelled.

type_

Type of this order document