Struct google_dfareporting2d4::PricingSchedule

pub struct PricingSchedule {
    pub start_date: Option<String>,
    pub end_date: Option<String>,
    pub flighted: Option<bool>,
    pub disregard_overdelivery: Option<bool>,
    pub cap_cost_option: Option<String>,
    pub pricing_periods: Option<Vec<PricingSchedulePricingPeriod>>,
    pub pricing_type: Option<String>,
    pub testing_start_date: Option<String>,
    pub floodlight_activity_id: Option<String>,
}

Pricing Schedule

This type is not used in any activity, and only used as part of another schema.

Fields
start_date: Option<String>

Placement start date. This date must be later than, or the same day as, the campaign start date. The hours, minutes, and seconds of the start date should not be set, as doing so will result in an error. This field is required on insertion.

end_date: Option<String>

Placement end date. This date must be later than, or the same day as, the placement start date, but not later than the campaign end date. If, for example, you set 6/25/2015 as both the start and end dates, the effective placement date is just that day only, 6/25/2015. The hours, minutes, and seconds of the end date should not be set, as doing so will result in an error. This field is required on insertion.

flighted: Option<bool>

Whether this placement is flighted. If true, pricing periods will be computed automatically.

disregard_overdelivery: Option<bool>

Whether cap costs are ignored by ad serving.

cap_cost_option: Option<String>

Placement cap cost option.

pricing_periods: Option<Vec<PricingSchedulePricingPeriod>>

Pricing periods for this placement.

pricing_type: Option<String>

Placement pricing type. This field is required on insertion.

testing_start_date: Option<String>

Testing start date of this placement. The hours, minutes, and seconds of the start date should not be set, as doing so will result in an error.

floodlight_activity_id: Option<String>

Floodlight activity ID associated with this placement. This field should be set when placement pricing type is set to PRICING_TYPE_CPA.