Struct google_dfareporting2d4::ReportFloodlightCriteriaReportProperties

pub struct ReportFloodlightCriteriaReportProperties {
    pub include_unattributed_ip_conversions: Option<bool>,
    pub include_unattributed_cookie_conversions: Option<bool>,
    pub include_attributed_ip_conversions: Option<bool>,
}

The properties of the report.

This type is not used in any activity, and only used as part of another schema.

Fields
include_unattributed_ip_conversions: Option<bool>

Include conversions that have no associated cookies and no exposures. It’s therefore impossible to know how the user was exposed to your ads during the lookback window prior to a conversion.

include_unattributed_cookie_conversions: Option<bool>

Include conversions of users with a DoubleClick cookie but without an exposure. That means the user did not click or see an ad from the advertiser within the Floodlight group, or that the interaction happened outside the lookback window.

include_attributed_ip_conversions: Option<bool>

Include conversions that have no cookie, but do have an exposure path.