Struct google_dfareporting2d4::ReportReachCriteria

pub struct ReportReachCriteria {
    pub activities: Option<Activities>,
    pub dimensions: Option<Vec<SortedDimension>>,
    pub metric_names: Option<Vec<String>>,
    pub date_range: Option<DateRange>,
    pub dimension_filters: Option<Vec<DimensionValue>>,
    pub reach_by_frequency_metric_names: Option<Vec<String>>,
    pub custom_rich_media_events: Option<CustomRichMediaEvents>,
    pub enable_all_dimension_combinations: Option<bool>,
}

The report criteria for a report of type "REACH".

This type is not used in any activity, and only used as part of another schema.

Fields

activities

Activity group.

dimensions

The list of dimensions the report should include.

metric_names

The list of names of metrics the report should include.

date_range

The date range this report should be run for.

dimension_filters

The list of filters on which dimensions are filtered. Filters for different dimensions are ANDed, filters for the same dimension are grouped together and ORed.

reach_by_frequency_metric_names

The list of names of Reach By Frequency metrics the report should include.

custom_rich_media_events

Custom Rich Media Events group.

enable_all_dimension_combinations

Whether to enable all reach dimension combinations in the report. Defaults to false. If enabled, the date range of the report should be within the last three months.