Struct google_dfareporting2d4::TagSetting

pub struct TagSetting {
    pub include_click_through_urls: Option<bool>,
    pub include_click_tracking: Option<bool>,
    pub additional_key_values: Option<String>,
    pub keyword_option: Option<String>,
}

Tag Settings

This type is not used in any activity, and only used as part of another schema.

Fields

include_click_through_urls

Whether static landing page URLs should be included in the tags. This setting applies only to placements.

include_click_tracking

Whether click-tracking string should be included in the tags.

additional_key_values

Additional key-values to be included in tags. Each key-value pair must be of the form key=value, and pairs must be separated by a semicolon (;). Keys and values must not contain commas. For example, id=2;color=red is a valid value for this field.

keyword_option

Option specifying how keywords are embedded in ad tags. This setting can be used to specify whether keyword placeholders are inserted in placement tags for this site. Publishers can then add keywords to those placeholders.